requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
Flask-Cors==3.0.10
PyPDF2==3.0.1
# fpdf==1.7.2
fpdf2
transformers==4.28.0
# sentencepiece==0.1.97
torch
scikit-learn==1.2.2
numpy==1.24.3
